change the type of a bunch of variables from `int' to `size_t'

+++ src/frontend/help/readhelp.c
@@ -42,7 +42,7 @@ static void
 sortlist(toplink **tlp)
 {
     toplink **vec, *tl;
-    int num = 0, i;
+    size_t num = 0, i;
 
     for (tl = *tlp; tl; tl = tl->next)
         num++;
