+++ src/spicelib/devices/numd/numdset.c
@@ -37,7 +37,7 @@ NUMDsetup(SMPmatrix *matrix, GENmodel *inModel, CKTcircuit *ckt, int *states)
   OPTNcard *options;
   OUTPcard *outputs;
   char *icFileName = NULL;
-  int nameLen;
+  size_t nameLen;
   int error;
   int xMeshSize;
   ONEdevice *pDevice;
